:sparkles: Begin TypeScript conversion
