+++ package.json
@@ -1,29 +1,45 @@
 {
-  "name": "mc-font-extractor",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "bin": {
-    "mc-font-extractor": "./cli.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Prouser123/mc-font-extractor.git"
-  },
-  "author": "jcx",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/Prouser123/mc-font-extractor/issues"
-  },
-  "homepage": "https://github.com/Prouser123/mc-font-extractor#readme",
-  "dependencies": {
-    "colors": "^1.4.0",
-    "fontx": "^1.3.0",
-    "jszip": "^3.5.0",
-    "node-fetch": "^2.6.1",
-    "opentype.js": "^1.3.3",
-    "prefix-logger": "0.0.2",
-    "upng-js": "^2.1.0",
-    "yargs": "^16.2.0"
-  }
+    "name": "mc-font-extractor",
+    "version": "1.0.0",
+    "description": "",
+    "main": "dist/index.js",
+    "bin": {
+        "mc-font-extractor": "./dist/cli.js"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/DumbDogDiner/mc-font-extractor.git"
+    },
+    "contributors": [
+        {
+            "name": "jcx"
+        },
+        {
+            "name": "ZacheryFox",
+            "email": "zachery@foxboy.sh"
+        },
+        {
+            "name": "SkyezerFox",
+            "email": "actuallyori@gmail.com"
+        }
+    ],
+    "license": "MIT",
+    "bugs": {
+        "url": "https://github.com/DumbDogDiner/mc-font-extractor/issues"
+    },
+    "homepage": "https://github.com/DumbDogDiner/mc-font-extractor#readme",
+    "dependencies": {
+        "colors": "^1.4.0",
+        "fontx": "^1.3.0",
+        "jszip": "^3.5.0",
+        "node-fetch": "^2.6.1",
+        "opentype.js": "^1.3.3",
+        "prefix-logger": "0.0.2",
+        "upng-js": "^2.1.0",
+        "yargs": "^16.2.0"
+    },
+    "devDependencies": {
+        "@types/node": "^14.14.22",
+        "@types/node-fetch": "^2.5.8"
+    }
 }
